Moved database to a folder under the BlogEngine application. Deleted old location. Modified ConfigManager.cs so that it handles a relative connection string. Added various other methods to DbUtil.cs. Changed int to long in entity classes as otherwise an error occurs. PostRepository.GetPosts now returns actual post data from the database.

+++ BlogEngine/Models/Entities/BaseEntity.cs
@@ -6,7 +6,7 @@ namespace BlogEngine.Models.Entities
     {
         public DateTime CreatedOn { get; set; }
         public DateTime ModifiedOn { get; set; }
-        public int CreatedBy { get; set; }
-        public int ModifiedBy { get; set; }
+        public long CreatedBy { get; set; }
+        public long ModifiedBy { get; set; }
     }
 }
